Check if default app protocol fix
This PR should fix the app always setting itself as the default protocol when it's not necessary to. On some operating systems (notably seen with linux) maybe rather has to do with snap in the referenced issue it closes it seems to always show a confirmation popup on every launch when the app is setting itself as a protocol client.